Moved POC elevator widget to normal_body_right; renamed HOC
